requirements.txt
fastapi[all]
pydantic
uvicorn
boto3
pillow
mangum
aws-cdk-lib
constructs
